requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
pyotp==2.9.0
pyserial==3.5
